app/requirements.txt
fastapi
httpx
sqlalchemy
pymysql
python-dotenv
uvicorn
PyJWT
cryptography